+++ locksmith/scripts/deploy-elasticbeanstalk-docker.sh
@@ -8,7 +8,7 @@ node_env=$7
 is_forked_pr=$8
 build_id=$9
 message=$10
-staging_environment=${environment}-staging
+stripe_secret=$11
 
 function check_is_forked_pr()
 {
@@ -36,9 +36,10 @@ function deploy()
     environment_name=$1
 
     if eb status ${environment_name}; then
+        eb setenv DB_USERNAME=${db_username} DB_PASSWORD=${db_password} DB_NAME=${db_name} DB_HOSTNAME=${db_hostname} NODE_ENV=${node_env} STRIPE_SECRET=${stripe_secret}
         eb deploy ${environment_name} --label locksmith-${build_id} --message "${message:0:199}"
     else
-        eb create ${environment_name} --envvars DB_USERNAME=${db_username},DB_PASSWORD=${db_password},DB_NAME=${db_name},DB_HOSTNAME=${db_hostname},NODE_ENV=${node_env} --elb-type classic
+        eb create ${environment_name} --envvars DB_USERNAME=${db_username},DB_PASSWORD=${db_password},DB_NAME=${db_name},DB_HOSTNAME=${db_hostname},NODE_ENV=${node_env},STRIPE_SECRET=${stripe_secret} --elb-type classic
     fi
     
 }