+++ paywall/src/__tests__/data-iframe/start/connectToBlockchain.test.js
@@ -113,6 +113,7 @@ describe('connectToBlockchain', () => {
           requiredConfirmations,
           update: onChange,
           window,
+          locksmithHost,
         })
       )
     })