+++ paywall/src/data-iframe/start/connectToBlockchain.js
@@ -57,16 +57,20 @@ export default async function connectToBlockchain({
     readOnlyProvider,
     blockTime,
     requiredConfirmations,
+    window,
+    onChange,
+    locksmithHost,
   })
 
   // sets up key purchase as available
   setPurchaseKeyCallback(
     makePurchaseCallback({
+      window,
+      locksmithHost,
       walletService,
       web3Service,
       requiredConfirmations,
       update: onChange,
-      window,
     })
   )