+++ paywall/src/data-iframe/start/makePurchaseCallback.js
@@ -15,6 +15,7 @@ const makePurchaseCallback = ({
   requiredConfirmations,
   update,
   window,
+  locksmithHost,
 }) =>
   async function purchase(lockAddress /*, extraTip */) {
     // note that the dynamic import is what makes code splitting
@@ -36,26 +37,29 @@ const makePurchaseCallback = ({
     ])
     const startingKey = keys[lockAddress]
 
-    // if purchaseKey errors out, it will emit an error, which
-    // is listened for by processKeyPurchaseTransactions, aborting both
-    // promises
-    return Promise.all([
-      purchaseKey({
-        walletService,
-        lockAddress,
-        amountToSend: locks[lockAddress].keyPrice,
-      }),
-      processKeyPurchaseTransactions({
-        walletService,
-        web3Service,
-        startingTransactions,
-        startingKey,
-        lockAddress,
-        requiredConfirmations,
-        update,
-        walletAction: () => update({ walletAction: true }),
-      }),
-    ])
+    try {
+      await Promise.all([
+        purchaseKey({
+          walletService,
+          lockAddress,
+          amountToSend: locks[lockAddress].keyPrice,
+        }),
+        processKeyPurchaseTransactions({
+          walletService,
+          web3Service,
+          startingTransactions,
+          startingKey,
+          lockAddress,
+          requiredConfirmations,
+          update,
+          walletAction: () => update({ walletModal: true }),
+          window,
+          locksmithHost,
+        }),
+      ])
+    } catch (error) {
+      update({ error: error.message })
+    }
   }
 
 export default makePurchaseCallback