+++ paywall/src/providers/Web3ProxyProvider.js
@@ -1,9 +1,10 @@
-import { iframePostOffice, setHandler } from '../utils/postOffice'
+import { iframePostOffice } from '../utils/postOffice'
 import {
   POST_MESSAGE_WEB3,
-  POST_MESSAGE_READY,
+  POST_MESSAGE_READY_WEB3,
   POST_MESSAGE_WALLET_INFO,
 } from '../paywall-builder/constants'
+import { waitFor } from '../utils/promises'
 
 /**
  * an iframe->main window web3 wallet proxy provider
@@ -18,20 +19,28 @@ import {
 export default class Web3ProxyProvider {
   constructor(window) {
     // set up the post office inside the iframe
-    this.postMessage = iframePostOffice(window)
+    const { postMessage, addHandler } = iframePostOffice(
+      window,
+      'Web3ProxyProvider',
+      'main window (web3 proxy)'
+    )
+    this.postMessage = postMessage
+    this.waiting = true
 
     // this is the postMessage version of connecting to the wallet
-    setHandler(POST_MESSAGE_WALLET_INFO, walletInfo => {
+    addHandler(POST_MESSAGE_WALLET_INFO, walletInfo => {
       if (!walletInfo || typeof walletInfo !== 'object') {
         return
       }
       this.isMetamask = !!walletInfo.isMetamask
       this.noWallet = !!walletInfo.noWallet
       this.notEnabled = !!walletInfo.notEnabled
+      // now we are ready to do work
+      this.waiting = false
     })
 
     // this is the postMessage version of the callback for sendAsync
-    setHandler(POST_MESSAGE_WEB3, web3Result => {
+    addHandler(POST_MESSAGE_WEB3, web3Result => {
       if (
         !web3Result.hasOwnProperty('error') ||
         !web3Result.hasOwnProperty('result')
@@ -54,14 +63,16 @@ export default class Web3ProxyProvider {
     this.requests = {}
 
     this.noWallet = true // until we hear back from the main window
-    this.postMessage(POST_MESSAGE_READY)
+    this.postMessage(POST_MESSAGE_READY_WEB3)
   }
 
   /**
    * @param {object} methodCall this object contains the method and params for the json-rpc call
    * @param {function} callback this is a standard node callback
    */
-  sendAsync({ method, params }, callback) {
+  async sendAsync({ method, params }, callback) {
+    // don't do anything until we have heard back on the wallet status
+    await waitFor(() => !this.waiting)
     if (this.noWallet) {
       callback(new Error('no ethereum wallet is available'))
       return