+++ rover/src/entity/RegistryKey.ts
@@ -1,0 +1,7 @@
+import { Entity, PrimaryColumn } from "typeorm";
+
+@Entity()
+export class RegistryKey {
+  @PrimaryColumn()
+  address: string;
+}