+++ unlock-js/src/__tests__/v10/getLock.test.js
@@ -3,9 +3,11 @@ import { ethers } from 'ethers'
 import Web3Service from '../../web3Service'
 import NockHelper from '../helpers/nockHelper'
 import utils from '../../utils'
+import { ZERO } from '../../constants'
+import erc20 from '../../erc20'
 
 const endpoint = 'http://127.0.0.1:8545'
-const nock = new NockHelper(endpoint, false /** debug */, true /** ethers */)
+const nock = new NockHelper(endpoint, false /** debug */)
 let unlockAddress = '0xD8C88BE5e8EB88E38E6ff5cE186d764676012B0b'
 
 let web3Service
@@ -15,6 +17,14 @@ const requiredConfirmations = 12
 const lockAddress = '0xc43efe2c7116cb94d563b5a9d68f260ccc44256f'
 const checksumLockAddress = utils.toChecksumAddress(lockAddress)
 const owner = '0x90F8bf6A479f320ead074411a4B0e7944Ea8c9C1'
+const erc20ContractAddress = '0x89d24A6b4CcB1B6fAA2625fE562bDD9a23260359'
+
+jest.mock('../../erc20.js', () => {
+  return {
+    getErc20BalanceForAddress: jest.fn(() => Promise.resolve(0)),
+    getErc20Decimals: jest.fn(() => Promise.resolve(18)), // 18 is the most frequent default for ERC20
+  }
+})
 
 describe('v10', () => {
   describe('getLock', () => {
@@ -38,17 +48,16 @@ describe('v10', () => {
       await nock.resolveWhenAllNocksUsed()
     }
 
-    function callReadOnlyFunction({ maxKeys }) {
-      // check to see if this is v0 or v01
+    function nockGetLock({ maxKeys = 10, erc20ContractAddress = ZERO }) {
+      // check to see what version that is
       nock.ethGetCodeAndYield(
         lockAddress,
         UnlockV10.PublicLock.deployedBytecode
       )
-      // what version is this? v02 succeeds with version 1
       nock.ethCallAndYield(
         fakeContract.functions['publicLockVersion()'].encode([]),
         checksumLockAddress,
-        resultEncoder.encode(['uint256'], [2])
+        resultEncoder.encode(['uint256'], [3])
       )
 
       // retrieve the bytecode and compare to v01
@@ -104,16 +113,40 @@ describe('v10', () => {
         checksumLockAddress,
         resultEncoder.encode(['uint256'], [utils.toRpcResultNumber(17)])
       )
+
+      nock.ethCallAndYield(
+        contractMethods.tokenAddress.encode([]),
+        checksumLockAddress,
+        resultEncoder.encode(['uint256'], [erc20ContractAddress])
+      )
+
+      nock.ethCallAndYield(
+        contractMethods.publicLockVersion.encode([]),
+        checksumLockAddress,
+        resultEncoder.encode(['uint256'], [utils.toRpcResultNumber(3)])
+      )
     }
 
-    it('should trigger an event when it has been loaded woth an updated balance', async () => {
+    function nockGetEthLock({ maxKeys = 10, erc20ContractAddress = ZERO }) {
+      nockGetLock({ maxKeys, erc20ContractAddress })
+      nock.getBalanceForAccountAndYieldBalance(
+        lockAddress,
+        utils.toRpcResultNumber('0xdeadfeed')
+      )
+    }
+
+    function nockGetErc20Lock({ maxKeys = 10, erc20ContractAddress = ZERO }) {
+      nockGetLock({ maxKeys, erc20ContractAddress })
+    }
+
+    it('should trigger an event when it has been loaded with an updated balance', async () => {
       expect.assertions(2)
       await nockBeforeEach()
-      callReadOnlyFunction({ maxKeys: 10 })
+      nockGetEthLock({})
 
       web3Service.on('lock.updated', (address, update) => {
         expect(address).toBe(lockAddress)
-        expect(update).toMatchObject({
+        expect(update).toEqual({
           balance: utils.fromWei('3735944941', 'ether'),
           keyPrice: utils.fromWei('10000000000000000', 'ether'),
           expirationDuration: 2592000,
@@ -121,16 +154,54 @@ describe('v10', () => {
           owner,
           outstandingKeys: 17,
           asOf: 1337,
+          currencyContractAddress: null,
+          publicLockVersion: 3,
         })
       })
 
       await web3Service.getLock(lockAddress)
     })
 
+    it('should successfully yield a lock with an ERC20 currency, with the right balance', async () => {
+      expect.assertions(3)
+      await nockBeforeEach()
+      nockGetErc20Lock({ erc20ContractAddress })
+      const balance = 1929
+      erc20.getErc20BalanceForAddress = jest.fn(() => {
+        return Promise.resolve(balance)
+      })
+      const decimals = 3
+      erc20.getErc20Decimals = jest.fn(() => {
+        return Promise.resolve(decimals)
+      })
+
+      web3Service.on('lock.updated', (address, update) => {
+        expect(address).toBe(lockAddress)
+        expect(update).toEqual({
+          balance: '1.929',
+          keyPrice: utils.fromDecimal('10000000000000000', decimals),
+          expirationDuration: 2592000,
+          maxNumberOfKeys: 10,
+          owner,
+          outstandingKeys: 17,
+          asOf: 1337,
+          currencyContractAddress: erc20ContractAddress,
+          publicLockVersion: 3,
+        })
+      })
+
+      await web3Service.getLock(lockAddress)
+      expect(erc20.getErc20BalanceForAddress).toHaveBeenCalledWith(
+        erc20ContractAddress,
+        lockAddress,
+        web3Service.provider
+      )
+    })
+
     it('should successfully yield a lock with an unlimited number of keys', async () => {
       expect.assertions(3)
       await nockBeforeEach()
-      callReadOnlyFunction({
+      nockGetEthLock({
         maxKeys:
           '0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff',
       })
@@ -151,6 +222,8 @@ describe('v10', () => {
         maxNumberOfKeys: -1,
         outstandingKeys: 17,
         owner: '0x90F8bf6A479f320ead074411a4B0e7944Ea8c9C1',
+        currencyContractAddress: null,
+        publicLockVersion: 3,
       })
     })
   })